requirements.txt
Django==2.1
django-storages==1.7.1
boto3==1.9.191